Added most of IM4M handling
There is an issue with the IM4M handler which causes junk bytes to be printed
instead of actual data. This needs to be addressed and can be observed by running

./img4helper --print-all test/iBoot.test.img4

The IM4P is printed correctly however when printing a standard .im4p file like
so:

./img4helper --print-all test/iBoot.RELEASE.d22.im4p

In both the IM4M is garbage. This is probably a bug in the asn1.c print functions
or something im doing when handling the im4m/img4 files.