Can now recognise file type
img4helper will now check if the files header whether the file is an IM4P, IM4M,
IM4R or a full IMG4. This at the moment is used with the --print-all/-a flag to
verify the given images type, and then will print everything as needed. This
means that passing an .im4p file to --print-all will not treat it like an IMG4.

There is still issues with reading IM4Ms, partly to do with the fact i've done
something wrong somewhere and partly due to me not implementing it fully yet.