Fixed issue with the size of uncompressed lzss kernelcaches
There was an issue that when running -e on an LZSS Compressed Kernelcache,
the process would write the wrong uncompressed file size, therefore a number
of the segments of the Kernel Mach-O point to a file offset that is beyond
the end of the file. This was apparent during development of HTool, where
lzss-compressed kernelcaches would fail when trying to decompile them, same
when loading into IDA64.

The issue is that in the img4_decompress_lzss () function (img4.c), I had
missed the ntohl () function on the compressed and uncompressed sizes in
two locations.

The first being line 545, where I output the compressed and uncompressed
sizes from the lzss compression header. And the second being when setting
the __TEXT	__DATA	__OBJC	others	dec	hex property of the  being returned by the function.

Setting this size property wrong meant that when img4helper tries to write
that image to a file, it creates a file on  bytes, but  bytes is far
too small for the actual size of the uncompressed image.

I've also taken the time to update libhelper. Img4helper doesn't make much
use of libhelper yet, but i will replace Glib stuff with the replacments
from libhelper.