Apple Silicon (arm64-dtk) Support & others
A few changes have been made. Firstly all header files are now
placed in their own include directory, the -o and --outfile arg
is now implemented and actually works, so users can specify
an outfile to write results to. When decrypting an image we now
no longer specify that no compression type was found, instead
that we failed to detect a compression type.

Finally is the arm64 Apple Silicon support. This did not require
any changes, only the changes made to libhelper in the commit
from arm64-dtk branch.

A Beta version of Img4helper for ARM64 MacOS will be released
soon